Adapt postMessage to be a little more kotlin-friendly
